+++ packages/iocuak/src/metadata/decorators/injectable.ts
@@ -2,6 +2,7 @@ import { BindingType } from '../../binding/models/domain/BindingType';
 import { TypeBinding } from '../../binding/models/domain/TypeBinding';
 import { Newable } from '../../common/models/domain/Newable';
 import { ServiceId } from '../../common/models/domain/ServiceId';
+import { taskScopeApiToTaskScopeMap } from '../../task/models/api/taskScopeApiToTaskScopeMap';
 import { TaskScope } from '../../task/models/domain/TaskScope';
 import { InjectableOptionsApi } from '../models/api/InjectableOptionsApi';
 import { MetadataKey } from '../models/domain/MetadataKey';
@@ -14,7 +15,10 @@ export function injectable(bindingApi?: InjectableOptionsApi): ClassDecorator {
     const bindingId: ServiceId =
       bindingApi?.id ?? (target as unknown as Newable);
 
-    const bindingScope: TaskScope = bindingApi?.scope ?? TaskScope.transient;
+    const bindingScope: TaskScope =
+      bindingApi?.scope === undefined
+        ? TaskScope.transient
+        : taskScopeApiToTaskScopeMap[bindingApi.scope];
 
     const binding: TypeBinding = {
       bindingType: BindingType.type,