+++ packages/iocuak/src/metadata/models/api/InjectableOptionsApi.ts
@@ -1,7 +1,7 @@
 import { ServiceId } from '../../../common/models/domain/ServiceId';
-import { TaskScope } from '../../../task/models/domain/TaskScope';
+import { TaskScopeApi } from '../../../task/models/api/TaskScopeApi';
 
 export interface InjectableOptionsApi {
   id?: ServiceId;
-  scope?: TaskScope;
+  scope?: TaskScopeApi;
 }