+++ packages/iocuak/CHANGELOG.md
@@ -26,7 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `ContainerApiService` interface is now exposed.
 - [BC]: Updated `ContainerApi` constructor to be private.
 - [BC]: Renamed `BindingApi` to `InjectableOptionsApi`.
-
+- [BC]: Renamed `TaskScope` to `TaskScopeApi`.